requirements.txt
pynput>=1.7.6
pystray>=0.19.4
Pillow>=10.0.0
pyinstaller>=6.0.0
